+++ source/Nuke.CodeGeneration/Generators/StringExtensions.cs
@@ -7,9 +7,7 @@ using System.Diagnostics;
 using System.Globalization;
 using System.Linq;
 using Humanizer;
-using JetBrains.Annotations;
 using Nuke.Common;
-using Nuke.Common.Utilities;
 
 namespace Nuke.CodeGeneration.Generators
 {