+++ source/Nuke.Common/Execution/ExecutionPlanHtmlService.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Reflection;
 using System.Text;
 using Nuke.Common.Utilities;
 using Nuke.Common.Utilities.Collections;
@@ -66,7 +65,10 @@ namespace Nuke.Common.Execution
             var builder = new StringBuilder();
 
             // When not hovering anything, highlight the default plan
-            var defaultPlan = ExecutionPlanner.GetExecutionPlan(executableTargets, new[] { executableTargets.Single(x => x.IsDefault).Name });
+            var defaultTarget = executableTargets.SingleOrDefault(x => x.IsDefault);
+            var defaultPlan = defaultTarget != null
+                ? ExecutionPlanner.GetExecutionPlan(executableTargets, new[] { defaultTarget?.Name })
+                : new ExecutableTarget[0];
             defaultPlan.ForEach(x => builder.AppendLine($@"  $(""#{x.Name}"").addClass('highlight');"));
 
             foreach (var executableTarget in executableTargets)