+++ source/Nuke.Common/Execution/HandleShellCompletionAttribute.cs
@@ -6,14 +6,16 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Nuke.Common.IO;
-using Nuke.Common.Tooling;
 
 namespace Nuke.Common.Execution
 {
     [AttributeUsage(AttributeTargets.Class)]
     internal class HandleShellCompletionAttribute : Attribute, IPreLogoBuildExtension
     {
-        public void Execute(NukeBuild build, IReadOnlyCollection<ExecutableTarget> executableTargets)
+        public void Execute(
+            NukeBuild build,
+            IReadOnlyCollection<ExecutableTarget> executableTargets,
+            IReadOnlyCollection<ExecutableTarget> executionPlan)
         {
             var completionItems = new SortedDictionary<string, string[]>();