+++ source/Nuke.Common/Execution/ParameterService.cs
@@ -8,7 +8,6 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using JetBrains.Annotations;
-using Nuke.Common.Tooling;
 using Nuke.Common.Utilities;
 using Nuke.Common.Utilities.Collections;
 using static Nuke.Common.Execution.ReflectionService;
@@ -161,7 +160,7 @@ namespace Nuke.Common.Execution
                 return GetDefaultValue(destinationType);
 
             return ConvertCommandLineArguments(
-                $"$all-positional",
+                "$all-positional",
                 positionalArguments,
                 destinationType,
                 _commandLineArguments,
@@ -205,7 +204,7 @@ namespace Nuke.Common.Execution
         private int GetCommandLineArgumentIndex(string argumentName, bool checkNames)
         {
             var index = Array.FindLastIndex(_commandLineArguments,
-                x => x.StartsWith("-") && x.Replace("-", string.Empty).EqualsOrdinalIgnoreCase(argumentName));
+                x => x.StartsWith("-") && x.Replace("-", string.Empty).EqualsOrdinalIgnoreCase(argumentName.Replace("-", string.Empty)));
 
             if (index == -1 && checkNames)
             {