+++ source/Nuke.Common/Tooling/Configure.cs
@@ -6,7 +6,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using JetBrains.Annotations;
 using Nuke.Common.Utilities.Collections;