+++ source/Nuke.Common/Tooling/ToolSettings.When.cs
@@ -3,10 +3,7 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
-using System.Collections.Generic;
-using System.Diagnostics.CodeAnalysis;
 using System.Linq;
-using JetBrains.Annotations;
 
 namespace Nuke.Common.Tooling
 {