+++ source/Nuke.GlobalTool.Tests/ProgramTest.cs
@@ -3,10 +3,8 @@
 // https://github.com/nuke-build/nuke/blob/master/LICENSE
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
-using Nuke.Common;
 using Xunit;
 
 namespace Nuke.GlobalTool.Tests